Refactor modules and add NameResolutionBlock
