+++ Cargo.toml
@@ -15,3 +15,4 @@ categories = ["encoding", "parsing"]
 [dependencies]
 byteorder = "1.3.2"
 derive-into-owned = "0.1.0"
+failure = "0.1.5"
