+++ src/pcapng/blocks/simple_packet.rs
@@ -8,13 +8,14 @@ use std::borrow::Cow;
 
 /// The Simple Packet Block (SPB) is a lightweight container for storing the packets coming from the network.
 /// Its presence is optional.
+#[derive(Clone, Debug)]
 pub struct SimplePacketBlock<'a> {
 
     /// Actual length of the packet when it was transmitted on the network.
-    original_len: u32,
+    pub original_len: u32,
 
     /// The data coming from the network, including link-layer headers.
-    data: &'a [u8]
+    pub data: &'a [u8]
 }
 
 impl<'a> SimplePacketBlock<'a> {
